--- a/instrumentation/github.com/gorilla/mux/otelmux/internal/semconv/util.go
+++ b/instrumentation/github.com/gorilla/mux/otelmux/internal/semconv/util.go
@@ -78,7 +78,16 @@ func serverClientIP(xForwardedFor string) string {
 
 func netProtocol(proto string) (name string, version string) {
 	name, version, _ = strings.Cut(proto, "/")
-	name = strings.ToLower(name)
+	switch name {
+	case "HTTP":
+		name = "http"
+	case "QUIC":
+		name = "quic"
+	case "SPDY":
+		name = "spdy"
+	default:
+		name = strings.ToLower(name)
+	}
 	return name, version
 }