--- a/internal/shared/semconvutil/netconv.go.tmpl
+++ b/internal/shared/semconvutil/netconv.go.tmpl
@@ -200,6 +200,15 @@ func splitHostPort(hostport string) (host string, port int) {
 
 func netProtocol(proto string) (name string, version string) {
 	name, version, _ = strings.Cut(proto, "/")
-	name = strings.ToLower(name)
+	switch name {
+	case "HTTP":
+		name = "http"
+	case "QUIC":
+		name = "quic"
+	case "SPDY":
+		name = "spdy"
+	default:
+		name = strings.ToLower(name)
+	}
 	return name, version
 }