Good day,

While I was review some profiles I noticed that the open-telemetry code was calling `strings.ToLower` a bit more then I expected.
After some digging I notice that `netProtocol` was called multiple times causing `strings.ToLower` to be called multiple times.

Now for our application 99.9%  of all requests are `HTTP/1.1` (and I expect this is the same for many others) so instead of calling `strings.ToLower` every-time `netProtocol` now checks some well known protocols first.

Let me know what you think and thanks for reviewing this PR :smile:

P.S. sorry for not updating the changelog but I honestly have no idea how to document this change as it's internal.