@dmathieu 2025-02-26
Could you share benchmarks showing the actual impact of this change?